feat: add JinaBERT support, default to F16, and make embedding model configurable
Add JinaBERT (ALiBi) model backend alongside standard BERT, with auto-detection
from config.json. Switch default dtype from F32 to F16 for better performance on
Metal. Make `codemem init` resolve the embedding model from env/config instead of
hardcoding BAAI/bge-base-en-v1.5, and write a default config.toml when absent.

+++ crates/codemem-core/src/config.rs
@@ -189,7 +189,7 @@ pub struct EmbeddingConfig {
     pub cache_capacity: usize,
     /// Batch size for embedding forward passes (GPU memory trade-off).
     pub batch_size: usize,
-    /// Weight dtype: "f32" (default), "f16" (half precision), "bf16".
+    /// Weight dtype: "f16" (default, half precision), "f32" (full), "bf16".
     /// F16 halves memory and is faster on Metal GPU.
     pub dtype: String,
 }
@@ -203,7 +203,7 @@ impl Default for EmbeddingConfig {
             dimensions: 768,
             cache_capacity: 10_000,
             batch_size: 16,
-            dtype: "f32".to_string(),
+            dtype: "f16".to_string(),
         }
     }
 }
